Remove the broad U18 exemption and replace with a more precise one.
Advice from @billker that intention of the original exception was to
protect minors from the decisions of the parents or guardians. As they
can't enter into a contract we can't hold them to the regulations that
would have prevented them from second representation.

This amendment would localise that exception to the area where it has a
material effect - on their ability to reject a prior allegiance.

In this state it would treat everyone equally for the purposes of
eligibility and grant freedom to a minor entering adulthood the
opportunity to make their own decision.